Add checking if thumbnail exist

+++ app/libraries/ThumbnailGenerator.php
@@ -99,7 +99,8 @@ class ThumbnailGenerator {
 
         $pathname = $this->destination.$this->generateFileName($file).'_thumb.png';
 
-        unlink($pathname);
+        if(file_exists($pathname))
+            unlink($pathname);
         $im->writeImage($pathname);
         $im->destroy();
 
@@ -121,7 +122,8 @@ class ThumbnailGenerator {
 
         $pathname = $this->destination.$this->generateFileName($file).'_thumb.png';
 
-        unlink($pathname);
+        if(file_exists($pathname))
+            unlink($pathname);
         $im->writeImage($pathname);
         $im->destroy();
